Convert useEndpointAction
