+++ client/contexts/ServerContext/endpoints/v1/emojiCustom.ts
@@ -8,4 +8,7 @@ export type EmojiCustomEndpoints = {
 			};
 		};
 	};
+	'emoji-custom.delete': {
+		POST: (payload: { emojiId: ICustomEmojiDescriptor['_id'] }) => void;
+	};
 };
